fix: add space in grep command for better readability

+++ test/ci/pytest.sh
@@ -8,17 +8,15 @@
 
 set +e
 # set -x
-if [[ "$OSTYPE" == "darwin"* ]]; then
+
+if [[ "${OSTYPE}" == "darwin"* ]]; then
   TD_OS="Darwin"
 else
-  OS=$(cat /etc/*-release | grep "^NAME=" | cut -d= -f2)
-  len=$(echo ${#OS})
-  len=$((len - 2))
-  TD_OS=$(echo -ne ${OS:1:${len}} | cut -d" " -f1)
+  # Extract OS name safely
+  OS=$(grep "^NAME=" /etc/*-release | head -n1 | cut -d= -f2 | tr -d '"')
+  TD_OS=$(echo "$OS" | awk '{print $1}')
 fi
 
-UNAME_BIN=$(which uname)
-OS_TYPE=$($UNAME_BIN)
 export ASAN_OPTIONS=detect_odr_violation=0
 cd .
 
@@ -28,7 +26,7 @@ CODE_DIR=$(pwd)
 
 IN_TDINTERNAL="community"
 if [[ "$CODE_DIR" == *"$IN_TDINTERNAL"* ]]; then
-  cd ../..
+  cd ../../
 else
   cd ../
 fi
@@ -46,7 +44,6 @@ fi
 
 declare -x BUILD_DIR=$TOP_DIR/$BIN_DIR
 declare -x SIM_DIR=$TOP_DIR/sim
-PROGRAM=$BUILD_DIR/build/bin/tsim
 PRG_DIR=$SIM_DIR/tsim
 ASAN_DIR=$SIM_DIR/asan
 
@@ -64,20 +61,20 @@ if [ ${#SIM_DIR} -lt 10 ]; then
    exit 1
 fi
 
-rm -rf $SIM_DIR/*
+rm -rf "${SIM_DIR:?}/"*
 
-mkdir -p $PRG_DIR
-mkdir -p $ASAN_DIR
+mkdir -p "${PRG_DIR}"
+mkdir -p "${ASAN_DIR}"
 
-cd $CODE_DIR
+cd "$CODE_DIR" || exit
 ulimit -n 600000
 ulimit -c unlimited
 
 #sudo sysctl -w kernel.core_pattern=$TOP_DIR/core.%p.%e
-echo "ExcuteCmd:" $*
+echo "ExcuteCmd:" "$@"
 
 if [[ "$TD_OS" == "Alpine" ]]; then
-  $*
+  "$@"
 else
   AsanFile=$ASAN_DIR/psim.info
   echo "AsanFile:" $AsanFile
@@ -85,30 +82,31 @@ else
   unset LD_PRELOAD
   #export LD_PRELOAD=libasan.so.5
   #export LD_PRELOAD=$(gcc -print-file-name=libasan.so)
-  export LD_PRELOAD="$(realpath "$(gcc -print-file-name=libasan.so)") $(realpath "$(gcc -print-file-name=libstdc++.so)")"
+  LD_PRELOAD="$(realpath "$(gcc -print-file-name=libasan.so)") $(realpath "$(gcc -print-file-name=libstdc++.so)")"
+  export LD_PRELOAD
   echo "Preload AsanSo:" $?
 
-  $* -A 2>$AsanFile
+  "$@" -A 2>"$AsanFile"
 
 
   unset LD_PRELOAD
   for ((i = 1; i <= 20; i++)); do
-    AsanFileLen=$(cat $AsanFile | wc -l)
-    echo "AsanFileLen:" $AsanFileLen
-    if [ $AsanFileLen -gt 10 ]; then
+    AsanFileLen=$(cat "${AsanFile}" | wc -l)
+    echo "AsanFileLen:" "${AsanFileLen}"
+    if [ "$AsanFileLen" -gt 10 ]; then
       break
     fi
     sleep 1
   done
   # check case successful
-  AsanFileSuccessLen=$(grep -w -a"successfully executed" $AsanFile | wc -l)
-  echo "AsanFileSuccessLen:" $AsanFileSuccessLen
+  AsanFileSuccessLen=$(grep -w -a -c "successfully executed" "$AsanFile")
+  echo "AsanFileSuccessLen:" "$AsanFileSuccessLen"
 
   if [[ "$AsanFileSuccessLen" -gt 0 ]]; then
     echo "Execute script successfully and check asan"
     # TODO: to be refactored, need to check if taos* process is closed successfully
     sleep 3
-    $CODE_DIR/ci/checkAsan.sh
+    "$CODE_DIR"/ci/checkAsan.sh
   else
     echo "Execute script failure"
     exit 1
